* connected the weather api to the screens * passed fields values from fire step, plume step and metorology step to the calculation object * changed the wind direction do be double instead of int * refactored some names of the fire concentration class * added model types * added location to output result class * removed the infinite loop from weather manager * added some static methods to the controller * added current location to controller * added current weather to controller * added default wind data button to meteorology view

+++ HotSpotHealthCode/app/src/main/java/hotspothealthcode/BL/AtmosphericConcentration/AtmosphericConcentration.java
@@ -32,7 +32,7 @@ public class AtmosphericConcentration
     protected MeteorologicalConditions meteorologicalCondition;
     protected double referenceHeight;
     protected double windSpeedAtReferenceHeight;
-    protected int windDirection;
+    protected double windDirection;
     protected double surfaceRoughnessHeight; // cm
     protected int sampleTime;
     protected TerrainType terrainType;
@@ -63,7 +63,7 @@ public class AtmosphericConcentration
         this.windSpeedAtReferenceHeight = windSpeedAtReferenceHeight;
     }
 
-    public void setWindDirection(int windDirection) {
+    public void setWindDirection(double windDirection) {
         this.windDirection = windDirection;
     }
